Added logger and refactored code
