+++ SOLID.Core/Contracts/ILogger.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace PlzSuperTool.Contracts
+{
+    public interface ILogger : IDisposable
+    {
+        void WriteLine(string line);
+    }
+}
