--- SOLID.Implementations/Implementations/ZipRepository.cs
@@ -1,131 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Net.Http;
-using System.Text.Json;
-using System.Web;
-using PlzSuperTool.Contracts;
-
-namespace PlzSuperTool.Implementations
-{
-    public sealed class ZipRepository : IZipSource
-    {
-        private readonly IDictionary<string, string[]> cache = new Dictionary<string, string[]>();
-
-        public string[] GetZipsFrom(bool online, string cityname)
-        {
-            using var logger = File.AppendText("log.txt");
-            logger.WriteLine(DateTime.Now.ToShortDateString() + " - " + DateTime.Now.ToShortTimeString() + " - GetZipsFrom: " + cityname);
-
-            IList<string> zips = new List<string>();
-
-            if (cache.ContainsKey(cityname))
-            {
-                return cache[cityname];
-            }
-
-            if (online != true)
-            {
-                foreach (var line in File.ReadAllLines("DE.txt"))
-                {
-                    var words = line.Split('\t');
-
-                    if (words[2].StartsWith(cityname, StringComparison.InvariantCultureIgnoreCase))
-                    {
-                        zips.Add(words[1]);
-                    }
-                }
-            }
-            else
-            {
-                string citynameUrlEncoded = HttpUtility.UrlEncode(cityname);
-                const string baseUrl = "https://public.opendatasoft.com/api/records/1.0/search/";
-                string query = $"?dataset=georef-germany-postleitzahl&q={citynameUrlEncoded}&facet=plz_name&facet=lan_name&facet=lan_code&rows=9999";
-
-                var httpClient = new HttpClient();
-                var stream = httpClient.GetStreamAsync(baseUrl + query).Result;
-
-                var plzResponse = JsonSerializer.DeserializeAsync<PlzResponse>(stream).Result;
-
-                plzResponse.records.ToList().ForEach(record =>
-                {
-                    zips.Add(record.fields.plz_code);
-                });
-            }
-
-            cache.Add(cityname, zips.ToArray());
-            logger.WriteLine(DateTime.Now.ToShortDateString() + " - " + DateTime.Now.ToShortTimeString() + " - GetZipsFrom: " + cityname + " - " + zips.Count + " results");
-
-            return zips.ToArray();
-        }
-    }
-
-    public class PlzResponse
-    {
-        public int nhits { get; set; }
-        public Parameters parameters { get; set; }
-        public Record[] records { get; set; }
-        public Facet_Groups[] facet_groups { get; set; }
-    }
-
-    public class Parameters
-    {
-        public string dataset { get; set; }
-        public string q { get; set; }
-        public int rows { get; set; }
-        public int start { get; set; }
-        public string[] facet { get; set; }
-        public string format { get; set; }
-        public string timezone { get; set; }
-    }
-
-    public class Record
-    {
-        public string datasetid { get; set; }
-        public string recordid { get; set; }
-        public Fields fields { get; set; }
-        public Geometry1 geometry { get; set; }
-        public DateTime record_timestamp { get; set; }
-    }
-
-    public class Fields
-    {
-        public string krs_code { get; set; }
-        public string lan_code { get; set; }
-        public float[] geo_point_2d { get; set; }
-        public string plz_name { get; set; }
-        public string plz_name_long { get; set; }
-        public string lan_name { get; set; }
-        public string name { get; set; }
-        public string plz_code { get; set; }
-        public string krs_name { get; set; }
-        public Geometry geometry { get; set; }
-    }
-
-    public class Geometry
-    {
-        public float[][][] coordinates { get; set; }
-        public string type { get; set; }
-    }
-
-    public class Geometry1
-    {
-        public string type { get; set; }
-        public float[] coordinates { get; set; }
-    }
-
-    public class Facet_Groups
-    {
-        public string name { get; set; }
-        public Facet[] facets { get; set; }
-    }
-
-    public class Facet
-    {
-        public string name { get; set; }
-        public int count { get; set; }
-        public string state { get; set; }
-        public string path { get; set; }
-    }
-}
